What have been updated: 1. addShootRecord()
NoRepeatShoot() is now replaced by this method to add a new point to shootboard if the point has not been shot before. Else, a notification will be printed to ask player to try again.

2a. checkHit()

- Added a comment to respond to Ben's queries.

2b. Lives counter

- Added to the player automatically when the player object is created. Initial value is 5. Every time the player's ship is hit, lives counter minuses 1.

3. checkAllSunk() {New method}

- A new method for game to check if the player has lost the game. This method should be initiated only if checkHit() returns true.

*** Rachel's questions ***

1.
For checkHit(), who is responsible for updating ship's inSunk status if a ship is sunk? Current design assumes game is responsible.

If player is responsible, I suggest Ship class can change isSunk() to a method allowing easy switch to true rather than StrikePoint input.

2. Is it good to separate checkAllSunk() from checkHit? Current design assumes game is responsible for calling checkAllSunk() when checkHit() returns true.